requirements.txt
opencv-python
dlib
face_recognition
python-pam